# perf: list comprehension into tuple