I know this looks like a stupid change, but it faster.


`python -O test/external/external_test_speed_llama.py`
Master
```
codegen(0)      mean runtime:  163.41ms, runs:   226.75,  130.42,  153.60,  140.43,  165.87
codegen(1)      mean runtime:  151.51ms, runs:   186.51,  147.10,  134.03,  146.45,  143.48
methodcache     mean runtime:   95.64ms, runs:    96.14,   91.87,  105.24,   92.51,   92.43
profile         mean runtime:  288.83ms, runs:   294.96,  285.53,  286.74,  289.17,  287.78
```

This branch
```
codegen(0)      mean runtime:  146.92ms, runs:   202.84,  123.40,  128.18,  142.59,  137.59
codegen(1)      mean runtime:  140.57ms, runs:   182.68,  122.07,  125.13,  139.34,  133.60
methodcache     mean runtime:   89.00ms, runs:    90.11,   88.46,   88.54,   88.56,   89.34
profile         mean runtime:  279.01ms, runs:   293.13,  270.80,  275.18,  279.12,  276.84
```


`JIT=0 python3 -O examples/llama.py --gen 1 --prompt "Hello." --count 10 --temperature 0 --timing`

This branch
```
loaded weights in 2380.07 ms, 13.48 GB loaded at 5.66 GB/s
Hello.
enqueue in 330.22 ms
total 887.98 ms, 1.13 tok/s, 18.83 GB/s, param 15.18 GB/s
 I
enqueue in 182.48 ms
total 185.18 ms, 5.40 tok/s, 72.87 GB/s, param 72.79 GB/s
'
enqueue in 161.80 ms
total 165.91 ms, 6.03 tok/s, 81.34 GB/s, param 81.24 GB/s
m
enqueue in 147.75 ms
total 151.88 ms, 6.58 tok/s, 88.86 GB/s, param 88.75 GB/s
 a
enqueue in 147.57 ms
total 151.68 ms, 6.59 tok/s, 88.98 GB/s, param 88.87 GB/s
 
enqueue in 162.20 ms
total 166.34 ms, 6.01 tok/s, 81.14 GB/s, param 81.03 GB/s
2
enqueue in 148.70 ms
total 152.75 ms, 6.55 tok/s, 88.36 GB/s, param 88.24 GB/s
0
enqueue in 151.88 ms
total 156.05 ms, 6.41 tok/s, 86.50 GB/s, param 86.38 GB/s
 year
enqueue in 171.88 ms
total 175.97 ms, 5.68 tok/s, 76.71 GB/s, param 76.60 GB/s
 old
enqueue in 153.78 ms
total 157.88 ms, 6.33 tok/s, 85.50 GB/s, param 85.38 GB/s
 male
output validated
```


Master
```
loaded weights in 2422.06 ms, 13.48 GB loaded at 5.56 GB/s
Hello.
enqueue in 345.56 ms
total 982.70 ms, 1.02 tok/s, 17.02 GB/s, param 13.72 GB/s
 I
enqueue in 206.61 ms
total 209.24 ms, 4.78 tok/s, 64.49 GB/s, param 64.42 GB/s
'
enqueue in 153.88 ms
total 158.01 ms, 6.33 tok/s, 85.41 GB/s, param 85.31 GB/s
m
enqueue in 169.73 ms
total 173.81 ms, 5.75 tok/s, 77.64 GB/s, param 77.55 GB/s
 a
enqueue in 154.73 ms
total 158.84 ms, 6.30 tok/s, 84.97 GB/s, param 84.86 GB/s
 
enqueue in 171.65 ms
total 175.82 ms, 5.69 tok/s, 76.77 GB/s, param 76.66 GB/s
2
enqueue in 160.64 ms
total 164.85 ms, 6.07 tok/s, 81.87 GB/s, param 81.76 GB/s
0
enqueue in 158.37 ms
total 162.63 ms, 6.15 tok/s, 83.00 GB/s, param 82.88 GB/s
 year
enqueue in 175.09 ms
total 179.39 ms, 5.57 tok/s, 75.25 GB/s, param 75.14 GB/s
 old
enqueue in 163.71 ms
total 168.01 ms, 5.95 tok/s, 80.35 GB/s, param 80.23 GB/s
 male
output validated
```