@geohot 2024-06-08
This change has been discussed before. It's more complex and isn't true in all versions of Python. Won't merge.