@roelofvandijk 2024-09-13
There's a fast path for lists into tuples.
https://github.com/python/cpython/blob/main/Objects/abstract.c#L2012 credit: @tim-becker